requirements.txt
pillow
pyautogui
python-telegram-bot
selenium
speedtest-cli